requirements.txt
manim
numpy
scikit-image
imageio
pillow==9.5.0